requirements.txt
qrcode==7.4.2
Pillow==10.4.0
python-dotenv==1.0.1
loguru==0.7.2
